Add dot product for 1D arrays
Previously, the RPN calculator had no support for matrix operations.
Functionality for the dot product of two 1D arrays has been added.
These arrays must be a list of comma-separated numbers with no spaces.
The dot operator is specified in the calculator with the '@' symbol.
Finally, the Travis CI YAML file has been updated to factor in Python
code coverage with the "coverage" library.